Move flask_utils next to related module
